Cherry-pick 75b2d55d from 5.0 to alleviate stalls during PREPARE.
In at least one production deployment, with max_connections=1000, backends were
stalled in PREPARE phase of 2PC while checkpoints were occurring too frequently
(14 secs apart).  Checkpoints are expected to take longer after the fix
dc9a76e9.  The lock "TemporarySerializePreparesLock adds to the misery by
letting only one backend do the PREPARE phase.  The cherry-pick removes this
restriction by removing TemporarySerializePreparesLock, thereby alleviating the
effects of slower checkpoints.

Original commit message:

Fix race condition while preparing transaction instead of serializing prepares.

Removed the locking and some more cleanups.
Avoid looping again in FinishPrepared Transaction. prepare_lsn to commit
the transaction can be found using gxact which we have locked, seems
pointless to loop around again to scan the list.

This is modified patch for GPDB based on postgres patch:
https://github.com/postgres/postgres/commit/bb38fb0d43c8d7ff54072bfd8bd63154e536b384#diff-3ed77c70e54e7f56eff48f6157aba91e
Original Patch commit message:
To lock a prepared transaction's shared memory entry, we used to mark it
with the XID of the backend. When the XID was no longer active according
to the proc array, the entry was implicitly considered as not locked
anymore. However, when preparing a transaction, the backend's proc array
entry was cleared before transfering the locks (and some other state) to
the prepared transaction's dummy PGPROC entry, so there was a window where
another backend could finish the transaction before it was in fact fully
prepared.

To fix, rewrite the locking mechanism of global transaction entries. Instead
of an XID, just have simple locked-or-not flag in each entry (we store the
locking backend's backend id rather than a simple boolean, but that's just
for debugging purposes). The backend is responsible for explicitly unlocking
the entry, and to make sure that that happens, install a callback to unlock
it on abort or process exit.