Analyzedb should analyze root partitions [#140458013]
This is a backport from 5X_STABLE (0d924f2c6 and 92fcd8ae8b)

When analyzedb is run without --skip_root_partition flag like below,
it should also analyze the root partitions.

`analyzedb -d <dbname> -s <schema> --full -a`

The expected key format for `leaf_root_dict` used for the check
is a tuple form and the result returned from the SQL was a string
concatenated with ".". Because of this, candidates could not be matched
in the dictionary and we never got the root partition details.

Made the appropriate changes in analyzedb so that it will
identify and analyze on the root partitions.

Signed-off-by: Jemish Patel <jpatel-pivotal@pivotal.io>