Unassign pids when destroy a cgroup dir.
On dropping a resgroup the cgroup dir also will be removed, however if
there is pids assigned to this group then the dir can't be removed.

So the pids must be first unassigned before removal, however we might
not have permission to assign them back to the system toplevel cgroup.

As these pids belong to idle gangs, so we could reassign them to the
gpdb toplevel cgroup.

One other thing to care about is that all the QD & QEs on the same host
will attempt to do the removal, this would cause race condition issue.
So we use flock to allow only the one of them to do the job.