Readers shouldn't check lock waitMask if writer holds the lock
Otherwise there is a possibility of distributed deadlock.  One such deadlock is
caused by ENTRY_DB_SINGLETON reader entering LockAcquire when QD writer of the
same MPP session already holds the lock.  A backend from another MPP session is
already waiting on the lock with a lockmode that conflicts with the reader's
requested lockmode.  This results in waitMask conflict and the reader is
enqueued in the wait queue.  But the QD writer is never going to release the
lock because it's waiting for tuples from segments (QE writers/readers).  And
the QE writers/readers are also waiting for the ENTRY_DB_SINGLETON reader,
completing the cycle necessary for deadlock.

The fix is to avoid checking waitMask conflicts for a reader if writer of the
same MPP session already holds the lock.  In such a case the reader is granted
the lock as long as it does not conflict with existing holders of the lock.

Two isloation2 tests are added.  One simulates the above mentioned deadlock and
fails if it occurs.  Another ensures that granting locks to readers without
checking waitMask conflict does not starve existing waiters.

cf. https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/OS1-ODIK0P4/ZIzayBbMBwAJ

Signed-off-by: Xin Zhang <xzhang@pivotal.io>

Conflicts:
    src/test/isolation2/isolation2_schedule
     - isolation2 schedules differ between 5X_STABLE and 4.3_STABLE