Add new resgroup operation API to get system memory size.
Add ResGroupOps_GetTotalMemory() to get system memory size, on Linux
system it gets the information from sysinfo() and /proc/ file system.

resgroup-ops-cgroup.c is renamed to resgroup-ops-linux.c as currently it
contains not only cgroup but also linux specific operations.

Some internal functions in the file are also refactored to be more generic.

Signed-off-by: Ning Yu <nyu@pivotal.io>