Fix an assertion failure in resgroup.
Suppose a resgroup already has a session assigned but no query running.
In another session (not assigned to this resgroup) drop this resgroup
and recreate it, then an assertion failure would be triggered.

The cause is that the old resgroup's struct in shm is reused for the new
resgroup, so on a QE who only owns a weak ref to this struct may find the
struct's resgroup id different than its expection.

To fix it we add the checks for such cases.