resgroup: TBC: retire resWaiting in MyProc.
We can check whether it's in the resgroup wait queue to decide whether
it is waiting or not.

This change is to be double confirmed as it might be slower to check the
wait queue than checking a variable directly, and access a queue is not
atomic operation.

However the benefit to retire this variable is that we can avoid the
inconsistency between the queue and this variable.