resgroup: reduce redundant logic in AlterResourceGroup().
We used to use a switch block to check min and max boundaries of the
specific type, each case block was just copy pasted with minor changes,
by using the type limit desc array introduced in previous commits we can
simplify the logic and increase the code readability.

Due to the special max limit of concurrency there is still a ugly
redundant handling for it, maybe we can come back it later and find out
a proper solution.