TINC: sort when listing files and adding tests sqls
By default, we use os.listdir to grab the list of files and directories
to figure out the order of which test to run first. The expectation that
the sql files are run in a sorted order is due to the "consistent"
behavior that the machines we've been using when running our tinc tests.
Now with terraform, the file system index can be different and may list
out the directories out of order.

Sort the directories or files to have a deterministic setup on TINC.

Signed-off-by: Shoaib Lari <slari@pivotal.io>