Make sure to check greenplum_path.sh when installing a gppkg
Sometimes we add environment variables into greenplum_path.sh when
installing a gppkg, ensure that we're checking that instead of checking
the os environment when the test started.

You can hit an issue where if you have a fresh greenplum cluster and run
the gppkg test, then the required environments are not set yet and could
cause the tests to be skipped.