Improve stability of some flaky resource group test cases.
In resource group test cases we verify that pending queries can be waken
up when free slot and/or memory quota is released by a finished query,
we use the isolation2 test framework to quit an old query and use
pg_stat_activity to check the pending status.

One problem is that the 'q' flag provided by the test framework only
ensure that the connection is closed, however the backends can still be
alive for a short time, if we query pg_stat_activity quick enough then
we may see the old query, which is expected to quit already, appears in
the output of pg_stat_activity. This doesn's always happen, but it does
happen now and then and makes the tests flaky.

As a workaround we put a short sleep in such cases to improve the
stability of these test cases.