Do not create a join with Dummy Rels
In case of UPDATE/DELETE commands on a partitioned table as shown below,
planner produces an inferior plan.

```
CREATE TABLE foo( dt date, m character varying(5), dk character varying(32))
DISTRIBUTED BY (dk);

CREATE TABLE bar( dt date, m character varying(5), dk character varying(32))
DISTRIBUTED BY (dk) PARTITION BY LIST(dt) (PARTITION p1 VALUES('2017-09-17'),
PARTITION p2 VALUES('2017-09-03'), PARTITION p3 VALUES('2017-09-04'));

EXPLAIN DELETE FROM bar b WHERE b.dt = DATE '20170917'AND EXISTS ( SELECT 1
FROM foo f WHERE f.dt = DATE '20170904' -1 AND b.m = f.m);
```

In the above query, the target relation (bar) is a partitioned table and there
exists a predicate on partitioning key(dt) of bar which gives planner an
opportunity to do partition elimination.  However even after eliminating the
appropriate partitions, planner created dummy join nodes by joining them with
foo.  Below is the simplified version of the planner produced plan:

```
Delete
   -> Append
          -> Explicit Redistribute Motion
                -> HashAgg
                    -> Hash Join
                        Hash Cond: b.m = f.m
                        -> Redistribute Motion
                            -> Seq Scan foo f
                        -> Redistribute Motion
                            -> Seq Scan bar_1_prt_p1 b
          -> Explicit Redistribute Motion
                -> HashAgg
                    -> Hash Join
                        Hash Cond: b.m = f.m
                        -> Redistribute Motion
                            -> Seq Scan foo f
                        -> One-Time Filter: false
          -> Explicit Redistribute Motion
                ->HashAgg
                  -> Hash Join
                      Hash Cond: b.m = f.m
                      -> Redistribute Motion
                          -> Seq Scan foo f
                      -> One-Time Filter: false
```

The second and third hash join nodes are absolutely unnecessary and they do not
perform any meaningful operation. This plan is inferior and creates unnecessary
slices.  As reported in issue #151196989, the table has 40 partitions and we
generated a plan similar to above which resulted in 80 slices. This eventually
led to poor performance of the query.

If the target relation for UPDATE/DELETE is a partitioned table then we enter
`inheritance_planner()` wherein we traverse through the `append_rel_list` that
contains all the child partitions of bar.

For each child partition, we call `grouping_planner()` to perform the planning
for joining following rels:
1. bar_1_prt_p1, foo               => Hash Join #1 above
2. bar_1_prt_p2, foo               => Hash Join #2 above
3. bar_1_prt_p3, foo               => Hash Join #3 above

Now when we construct the rels for these partitions, we attach the predicate
`b.dt = DATE '20170917'` from the query as rel's `baserestrictinfo`.  Now in
`set_plain_rel_pathlist()`, we check if this rel can be excluded via constraint
exclusion. This happens in `relation_excluded_by_constraints()`.  Essentially
we retrieve the check constraint on the partition and check if they are refuted
by their `baserestrictinfo`.  For instance, in case of bar_1_prt_p2, the check
constraint is `dt = '2017-09-03'` and `baserestrictinfo` is dt = `2017-09-17`.
Hence this partition gets excluded and a dummy rel is created for this
partition.

However later in make_join_rel(), we join this dummy rel with foo. And thus we
see a Hash Join node with a one-time false filter as one of its children.  We
should not create a join if one of its children is dummy.

In 5.X_STABLE , we do check for dummy rels and skip creating a join when one or
both(depending on the jointype) of its children are dummy; hence we don't see
this issue in 5.X_STABLE.

Similar to 5.X_STABLE, we have added a dummy rel check in `make_join_rel()` so
that the `joinrel` does not get created with dummy children. Now we see a good
plan where we exclude bar's p2 and p3 partitions :

```
Delete
  -> Explicit Redistribute Motion
          -> Hash EXISTS Join
              Hash Cond: b.m = f.m
              -> Redistribute Motion
                  -> Seq Scan on bar_1_prt_p1 b
              -> Redistribute Motion
                  -> Seq Scan on foo_1_prt_p2 f
```

This commit also adds relevant test in partition_pruning.sql and fixes
ICG test in subselect wherein the issue is easily evident by looking at
the old plan in `subselect.out`.

Signed-off-by: Jemish Patel <jpatel@pivotal.io>