resgroup: do not error out in transaction callbacks.
ResGroupHashFind() and ResGroupHashRemove() will error out when the
resgroup cannot be found, but in transaction callback functions no error
should happen. So we added a parameter `raise` to these functions to
let the callers to control whether to error out on failure.

Also put other operations in try catch block in these callback functions
for safe.