resgroup: workaround a race condition on slot.
A slot maybe reused on QE while its still using by another QE.

To workaround the issue we change the free slot pool from a stack into a
queue, slots are allocated from the head and freed to the tail, we also
increased the pool size so a slot's reuse is postponed.

This is not a fix, just a quick workaround.