Modify criteria for dumping procedural languages
When including or excluding schemas in gpcrondump, functions may be
dumped that require a procedural language to be created upon restore.
However, languages are not part of any schema, so they are not already
included in such a filter.

This commit changes the behavior of gp_dump to always dump procedural
languages when including and excluding schemas.  It still will not
dump them when including or excluding tables.

Signed-off-by: Karen Huddleston <khuddleston@pivotal.io>