resgroup: reset resgroup before quitting a session in tests.
In resgroup tests we need to query pg_stat_activity to verify that
pending queries got waken up properly after quit of an old session.

We use the 'q' command of the isolation2 test framework to quit an old
session, however what it actually does is only disconnecting the client
connection, the underlying QD might still be alive. If we then query
pg_stat_activity quick enough, then this __exited__ session may still be
listed in the output, this lead to flaky result of these resgroup tests.

As a workaround we reset the resgroup before the 'q' commands, so even
if their QDs are still alive they will not appear in the
pg_stat_activity output as we have filters on resgroup names.