resgroup: clean up some uses of write_log().
If we will disable memory limit check after a call to write_log() then
we could reorder the code to disable it first then we could use elog()
or LOG_RESGROUP_DEBUG() instead of write_log().

If memory limit check is not to be disabled then write_log() is still
used.