Fix timezone of co_alter_table_add test
The answer files are generated with `US/Pacific` timezone in mind.
However, the test didn't fixed the timezone to `US/Pacific`, hence it
will fail if the host is running in a different timezone.

This fix use the GUC `timezone` to set the timezone for the test always
at `US/Pacific`.