Enforce Constraints in ORCA
ORCA version 2.48.0 has a new configuration parameter
`fEnforceConstraintsOnDML`, which enables/disables the enforcement of
constraints in ORCA.  If set to true, ORCA will add Assert nodes to
enforce constraints.  If set to false, ORCA will not add Assert nodes
and will leave it to the executor to enforce them.

For 4.3_STABLE, we set it to true which means we will have the same
behaviour as before where ORCA will continue to enforce constraints.
Setting it to false changes the user visible error messages and we do
not want this to impact customer outputs.