resgroup: do not call open() with O_DIRECTORY.
O_DIRECTORY is to avoid denial-of-service problems if opendir(3) is
called on a FIFO or tape device. But it might be defined on legacy
linux systems.

In our use case this DOS problem should not happen as the path to open
is just a hard coded local path in some sense.

So to compile on legacy linux systems we retire the use of this flag.