gpstop: speed up WorkerPool.print_progress
Instead of sleeping for ten seconds, wait for the work_queue to finish
with a timeout of ten seconds (this way we return immediately once the
queue has finished). This speeds up gpstop by about fifteen seconds on
my machine.

This solution uses a workaround to Queue.join()'s lack of timeout; see
	https://bugs.python.org/issue9634

Signed-off-by: C.J. Jameson <cjameson@pivotal.io>