Add bool variable validate_reloptions to OpenIntoRel
When the gp_default_storage_options GUC was set and a table was
rewritten with ALTER TABLE REORGANIZE, the temp table generated would
take the storage options of the GUC instead of from the original
table. We should not be looking at default storage options when doing
internal queries. To prevent this, we add a boolean variable called
validate_reloptions in OpenIntoRel to check if the intended
destination is DestIntoRel (internal query) to decide whether or not
to use the default storage options for the temp table.

Authors: Jimmy Yih and Abhijit Subramanya

This deviates from Greenplum 5.1+ which has a data structure called
QueryDispatchDesc. Since we do not have this in 4.3, it should be fine
to be direct about checking destination of DestIntoRel in OpenIntoRel
instead of indirectly referencing it through a boolean inside the
query dispatch descriptor. This commit is functionally the same as the
referenced 5.1 commits below.

Backport Reference:
https://github.com/greenplum-db/gpdb/commit/919bdf0ee60b7a5a7769232ec7cda81b417adaf0
https://github.com/greenplum-db/gpdb/commit/cb10c6fbec16eac46959c66f451b991b7b62eaf6