resgroup: retire MyProc->resWaiting.
resWaiting is only true when proc is in the wait queue, and is only
false when proc is not in the wait queue, so it's a redundant flag.

And checking proc's queuing status can be performed automatically,
so resWaiting can be safely retired.

Their helper functions procIsWaiting() and procIsInWaitQueue() are also
merged into one.