Make the default user gpadmin for run_tinc task
The current user of the container is 'root'. This does not work for
ssh'ing into CCP AWS clusters because 'root' is explicitly disabled for
ssh. This is a standard pattern across any AWS AMI.

This is a quick fix to unblock the pipeline. Some refactors may follow
this that change the run_tinc PRETEST_SCRIPT pattern.

cherry-picked from greenplum-db/gpdb.git 5408ef + e5fe076bcc