resgroup: remove a risky assert().
This assert() is on a slot that we just released, however before the
assert() we have attempt to wake up other groups, so it's possible that
the slot is already reused in the other group, so the assert() just
cause a crash.

Removed this assert() as we should not touch a released slot in any
case.