resgroup: also update memQuotaUsed on QEs.
This value is necessary on ALTER RESOURCE GROUP to decide the new memory
capabilities, but it was not updated on QEs in the past, so the memory
quota might be over released on QEs. This does not lead to runtime
errors, and the memory quota can be regained laterly, but the
performance can be affected in worst case.

To fix it now we update this value also on QEs.