Simplify chown in common CI task script for gpadmin setup
In previous iterations while porting the CI pipelines to higher
concourse versions it was necessary in concourse version 3.3.3 to add
sleep and retry logic to a chown that was failing. It was most likely
failing due to underlying issues with concourse and the overlay FS at
that verison.

In later versions, namely the current running verison of 3.4.1 this
logic is not neccessary. However it is still a known issue that
recursive chowns of suffieciently large directories take an absurd
amount of time.