Do not modify original RTE subquery in `set_subquery_pathlist()`
For the following UPDATE on a partitioned table:
UPDATE update_tab a SET col = a.id
FROM
(SELECT b.id FROM subq_tab1 b,
	(SELECT dt, subq_func(id::text, 'abcd') FROM subq_tab2) c
WHERE c.dt >= '2017-01-01' )d;

When planner pushed the filter "c.dt >= '2017-01-01'" inside the
subquery RTE (corresponding to subquery "c"), it directly modified the
root->parse->rtable.
This caused an issue while generating an APPEND plan wherein
inheritance_planner referred to the modified root->parse->rtable while
planning each append_rel.
For the above query, during processing first append_rel in inheritance_planner(), the
subquery RTE in original root->parse got modified with a new qual as :
{OPEXPR :opno 1098 :opfuncid 1090 :opresulttype 16 :opretset false :args ({VAR :varno 3 :varattno 1 :vartype 1082 :vartypmod -1 :varlevelsup 0 :varnoold 3 :varoattno 1} {CONST :consttype 1082 :constlen 4 :constbyval true :constisnull false :constvalue 4 [ 66 24 0 0 0 0 0 0 ]})}

For the subsequent rel's in append list, generated plan was based on the
modified root->parse, resulting into a wrong final qual for relation
bar; which not only was redundant but was also in an incorrect format.
For instance, in the following plan the scan filter for bar appearing in
the second child of Append, is of the form list of list:
```
Update (slice0; segments: 3)  (rows=31134 width=18)
   ->  Append  (cost=1.05..12045.75 rows=31134 width=18)
         ->  Explicit Redistribute Motion 3:3  (slice3; segments: 3)  (cost=1.05..6022.88 rows=15567 width=18)
               ->  Nested Loop  (cost=1.05..6022.88 rows=15567 width=18)
                     ->  Nested Loop  (cost=0.00..3103.07 rows=15567 width=18)
                           ->  Broadcast Motion 3:3  (slice1; segments: 3)  (cost=0.00..1.07 rows=1 width=0)
                                 ->  Subquery Scan c  (cost=0.00..1.03 rows=1 width=0)
                                       ->  Seq Scan on bar  (cost=0.00..1.02 rows=1 width=8)
                                             Filter: dt >= '2017-01-01'::date
                           ->  Seq Scan on update_tab_1_prt_def a  (cost=0.00..567.00 rows=15567 width=18)
                     ->  Materialize  (cost=1.05..1.08 rows=1 width=0)
                           ->  Broadcast Motion 3:3  (slice2; segments: 3)  (cost=0.00..1.05 rows=1 width=0)
                                 ->  Seq Scan on foo b  (cost=0.00..1.01 rows=1 width=0)
         ->  Explicit Redistribute Motion 3:3  (slice6; segments: 3)  (cost=1.05..6022.88 rows=15567 width=18)
               ->  Nested Loop  (cost=1.05..6022.88 rows=15567 width=18)
                     ->  Nested Loop  (cost=0.00..3103.07 rows=15567 width=18)
                           ->  Broadcast Motion 3:3  (slice4; segments: 3)  (cost=0.00..1.07 rows=1 width=0)
                                 ->  Subquery Scan c  (cost=0.00..1.03 rows=1 width=0)
                                       ->  Seq Scan on bar  (cost=0.00..1.02 rows=1 width=8)
                                             Filter: (dt >= '2017-01-01'::date) AND dt >= '2017-01-01'::date  -----> incorrect qual
                           ->  Seq Scan on update_tab_1_prt_2 a  (cost=0.00..567.00 rows=15567 width=18)
                     ->  Materialize  (cost=1.05..1.08 rows=1 width=0)
                           ->  Broadcast Motion 3:3  (slice5; segments: 3)  (cost=0.00..1.05 rows=1 width=0)
                                 ->  Seq Scan on foo b  (cost=0.00..1.01 rows=1 width=0)
```

This resulted in a crash in ExecQual.c since executor expects the qual
to be a list of Exprs such as OpExpr, DistinctExpr; however instead it
received a qual as a list of list.

This commit fixes the problem by making a modifiable copy of RTE
subquery in `set_subquery_pathlist()` before we act on it.

Added relevant test.