Retry 2PC incase of failure for vacuum.
vacuum starts distributed transaction for every relation it vacuums and also
commits them separately performing 2PC. But in this case if first attempt to
CommitPrepared fails, normally it gets retried via `EndCommand()` but in this
case we just move on to process next relation and miss really completing the
transaction fully.

Note:
This problem is not seen for 5.0 and later as logic to perform retries for 2PC
has been moved to `doNotifyingCommitPrepared()` itself instead of
`EndCommand()`. Hence point fixing this only for 4.3 by explicitely calling
`doDtxPhase2Retry()`.