resgroup: retry proc migration for rmdir to succeed.
To remove a cgroup dir we have to first migrate all its processes to
other cgroups, this operation can return before the underlying migration
is actually done, so a following rmdir operation can still fail if it's
quick enough.

To fix it we have to retry the migration and the rmdir operation.