resgroup: add delay in a testcase.
The 'q' command in the isolation2 test framework returns before the
underlying segments actually quit, so a following DROP comand might fail
if it's quick enough.

In the resgroup_alter_concurrency testcase there is a batch of 5 'q'
commands which can trigger this race condition easily on ubuntu system.

A delay is added as a quick workaround, as the root fix would need to
make the master waiting for ack from all the segments on quit which is
not a trivial change.