Fix pushing down of quals below subqueries with set operations
As part of the commit b8002a9, the check for
`qual_is_pushdown_safe_set_operation()` was removed and a check was
added in `subquery_is_pushdown_safe()` to bail out from pushing down
qual if the subquery->hasFuncs was true. However, the check was
too restrictive and prohibited pushdown of those quals which do not
contain any window reference resulting in performance regression.

This commit performs 2 things:
1. Revert the commit b8002a9
2. Fix `qual_is_pushdown_safe_set_operation` to correctly resolve the
qual vars and identify if there are any window references in the top
level of the set operation's left or right subqueries.  Before commit
b8002a9, instead of starting with rte of the level where the qual is
attached we started scanning the rte of the subqueries of the left and
right args of setop to identify the qual. However, because of this the
varno didn't match to the corresponding RTE due to which the quals
couldn't be resolved to winref and were incorrectly pushed down. This
caused the planner to return an error during execution.

Adding relevant tests.

Signed-off-by: Ekta Khanna <ekhanna@pivotal.io>