Ignore template0 from stop and warn limit calculations
Customers may already have scenarios where age of template0 is past warn limit
and stop limit. On binary swap to avoid restart failures ("database is not
accepting commands to avoid wraparound data loss in database") continue to skip
template0 from stop limit and warn limit calculations. The datfrozenxid of
template0 is still used to set the xidVacLimit used for triggering autovacuum.
And the datfrozenxid of template0 is also used for calculating transaction id
for clog truncation.

There are still scenarios where autovacuum may fail with this error "could not
access status of transaction XXX". This is because the clog was truncated
without considering template0. In order to recover from this scenario the
missing clog file must be manually created with transaction XXX in committed
state. It is very important to have the status of this missing transaction as
committed.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>