Fix ordering issues in gppkg TiNC Tests
After we migrated to GCP, the order of tests changed.  This change
highlighted the less than ideal isolation between TiNC test suites.
Commit 7e78764 addressed some of the test pollution, but recently the
test order has changed again.

This commit chooses to make the test order explicit.

Why are we only addressing the symptoms and not the test pollution?  We
have already committed to migrating all of the TiNC test suites to
Behave, and so we are chosing to not invest time in improving what we
are signed up to migrate in the near future.

Co-authored-by: Jim Doty <jdoty@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>