gdd: alloc MyProcPort on stack.
We used to allocate MyProcPort with malloc() and does not check for the
result, it's reported to trigger a crash in gprecoverseg test:

	(gdb) bt
	0x00007f56b506794f in __strlen_sse42 () from /lib64/libc.so.6
	0x00000000009f7894 in write_message_to_server_log ()
	0x00000000009fc200 in send_message_to_server_log ()
	0x00000000009ffb85 in EmitErrorReport ()
	0x00000000009fccf5 in errfinish ()
	0x0000000000acd99f in FtsTestSegmentDBIsDown ()

Now we allocate it directly on stack.

We can not find out a way to construct a testcase for this issue, but
according to the report it should be covered in existing tests.