gdd: store name of super user on stack.
We used to store name of super user in a buffer allocated by strdup()
and did not check for the result, it's reported to trigger a crash in
gprecoverseg test.

Now we store it in a buffer on stack.

A fatal error will be triggered if no super user can be found.

We can not figure out a way to construct a testcase for this change, but
according to the report it should be covered in existing tests.