Do not translate printable filters of PartitionSelector (#4959)
Printable filters are used to produce an expression for the Partition
Selector node that is printed during EXPLAIN to give a hint to the
general nature of the filter used by the node. It is not used by the
executor in any way which actually uses levelEqExpressions,
levelExpressions & residualPredicate instead. These usually contain
completely different set of expressions such as PartBounExpr etc. which
are not printed during EXPLAIN.

Also, with dynamic partition elimination, the partition selector's
printable filter may contain VARs that are not in its subtree and
instead refer to a DynamicTableScan node on the other side of a Join.
This means that it becomes tricky to extract the correct printable
filter expression during DXL to PlStmt translation since that occurs
bottom-up.

Since it is misleading and sometimes incorrect, it's better to remove it
altogether.

Signed-off-by: Sambitesh Dash <sdash@pivotal.io>
(cherry picked from commit c38c72ff979bf4a7e618b5a19700df8cc6d67159)

Also update the bfv_partition optimizer answer file since we no longer
translate the printable filter which was causing the "bogus var" error.