Fix compilation and test issues introduced by parruleord fix
Compilation was broken because PG_INT16_MAX constant was
unavailable. It was introduced in master during 5.0 era with commit
https://github.com/greenplum-db/gpdb/commit/94eacb66bdf. I have
backported only the constant to fix the compilation error.

The TINC parruleord test fails because of output differences between
pg_regress and TINC. The new git executable somehow took the
parruleord test changes in pg_regress and placed them in the old TINC
parruleord test. This is technically correct but was unexpected that
the test outputs would differ (hard to remember since TINC tests have
not had a test change in quite a while).