Fix querying stats from largest child correctly
Previously, we were assuming the attnums for both child and root table
would be the same. In the scenario where we delete a column and then
split a partition would expose a behavior where the assumption is
invalid. Therefore, we cannot query the stats using the root table's
attnum for a specific attribute (column).

This commit fixes the issue by getting the attribute name using the
root's attnume and use it to acquire correct attnum for the largest leaf
partition.