resgroup: fallback to insert when updating caps.
Resource group capabilities could be missing in ALTER command, e.g.:

- create resgroup rg1 with v5.0 which does not support cpuset(cap=7);
- binary switch to v5.10 (suppose it supports cpuset);
- alter rg1's cpu_rate_limit, it will also `update` the cpuset cap;

Now as rg1 was created with v5.0 so there is no cap=7 row in the catalog
table pg_resgroupcapability, so the `update` operation will raise an
error as the expected tuple can not be found.

The proper behavior is to fallback to `insert` in such a case.

Test cases are not included as it is already covered by existing binary
swap test resgroup_current_3_queue.